Refactor the session cookie handling
Always set the session cookie regardless of whether a Registrant
has been saved in the db. This allows us to more easily check
whether a request originates in the "middle" of the workflow,
and redirect to / if it does.

This fixes the exception thrown when trying to access a
non-existent Registrant, as when making an initial request
to any other path than /.